requirements.txt
asgiref==3.6.0
autopep8==2.0.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-appconf==1.0.5
django-channels==0.7.0
django-cors-headers==3.14.0
django-dynamic-preferences==1.14.0
django-environ==0.10.0
django-filter==23.1
django-imagekit==4.1.0
django-redis==5.2.0
django-rest-framework==0.1.0
django-silk==5.0.3
djangorestframework==3.14.0
google-auth==2.16.3
google-auth-oauthlib==1.0.0
gprof2dot==2022.7.29
idna==3.4
oauthlib==3.2.2
persisting-theory==1.0
pilkit==2.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pytz==2023.2
redis==4.5.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
